+++ lib/scss_lint/linter/unnecessary_mantissa.rb
@@ -6,6 +6,7 @@ module SCSSLint
 
     def visit_script_string(node)
       return unless node.type == :identifier
+      return if node.value =~ /^'|"/
 
       node.value.scan(REAL_NUMBER_REGEX) do |number, integer, mantissa, units|
         if unnecessary_mantissa?(mantissa)